Add future finance calculator

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
-income = int(input("Enter your monthly income: "))
-expenses = int(input("Enter your total monthly expenses: "))
-savings = income - expenses
+income = input("Enter your monthly income: ")
+expenses = input("Enter your total monthly expenses: ")
+savings = float(income) - float(expenses)
 annual_savings = savings * 12 + (savings * 12 * 0.05)
 
 print("Your monthly savings are $" + str(savings))
